LayoutTests/fast/text/text-orientation-parse-competition.html

<!DOCTYPE html>
<html>
	<head>
	<meta charset="utf-8">
	<style>
		#test3 { -webkit-text-orientation: sideways; text-orientation: upright; }

		#test4 { text-orientation: sideways; -webkit-text-orientation: upright; }

		#test5 { -webkit-text-orientation: sideways; }
		#test5 { text-orientation: upright; }

		#test6 { text-orientation: sideways; }
		#test6 { -webkit-text-orientation: upright; }
	</style>
	<script src="../../resources/js-test-pre.js"></script>
	</head>
	<body>
		<div id="test1" style="-webkit-writing-mode: vertical-rl; -webkit-text-orientation: sideways; text-orientation: upright">你好ABC</div>
		<div id="test2" style="-webkit-writing-mode: vertical-rl; text-orientation: sideways; -webkit-text-orientation: upright">你好ABC</div>
		<div id="test3" style="-webkit-writing-mode: vertical-rl">你好ABC</div>
		<div id="test4" style="-webkit-writing-mode: vertical-rl">你好ABC</div>
		<div id="test5" style="-webkit-writing-mode: vertical-rl">你好ABC</div>
		<div id="test6" style="-webkit-writing-mode: vertical-rl">你好ABC</div>
		<script>
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test1')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test2')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test3')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test4')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test5')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test6')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test1')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test2')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test3')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test4')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test5')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test6')).textOrientation", "upright");
		</script>
		<script src="../../resources/js-test-post.js"></script>
	</body>
</html>
